removed legacy components, cleaned up, small fixes
